Enable running the tests without external calls
